requirements.txt
# Python 依赖包
PyYAML>=6.0

